+++ log4k-slf4j/src/jsMain/kotlin/saschpe/log4k/slf4j/SLF4JLogger.kt
@@ -7,6 +7,6 @@ import saschpe.log4k.Logger
 actual class SLF4JLogger : Logger() {
     private val logger = ConsoleLogger()
 
-    override fun print(level: Log.Level, tag: String, message: String?, throwable: Throwable?) =
+    actual override fun print(level: Log.Level, tag: String, message: String?, throwable: Throwable?) =
         logger.log(level, tag, message, throwable)
 }